requirements.txt
# Flask y extensiones principales
Flask==2.3.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0

# Base de datos PostgreSQL
psycopg2-binary==2.9.9

# Serialización con Marshmallow
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Herramientas adicionales
SQLAlchemy==2.0.23
Werkzeug==3.0.1
python-dotenv==1.0.0

# AWS SDK y Logging
boto3==1.34.8
watchtower==3.0.1

# Google Cloud
google-cloud-storage==2.13.0

# Utilidades
PyJWT==2.8.0
gunicorn==21.2.0
requests==2.31.0
